+++ src/components/analogIO/model.h
@@ -36,7 +36,7 @@ public:
   bool EncodeAnalogIOEvent(char *pin_name, float pin_value,
                            wippersnapper_sensor_SensorType read_type);
   bool EncodeAnalogIOEventVoltage(char *pin_name, float pin_value);
-  bool EncodeAnalogIOEventRaw(char *pin_name, int16_t pin_value);
+  bool EncodeAnalogIOEventRaw(char *pin_name, float pin_value);
   wippersnapper_analogio_AnalogIOEvent *GetAnalogIOEvent();
 
 private:
